pokedex-backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
python-dotenv
beautifulsoup4
requests
